feat(SFS Turbo): add SFS Turbo data-sources and resources support.
